# Improve path search in Kestrel parsing